requirements.txt
discord.py
python-dotenv
requests
os
random
google-generativeai